new commited

+++ ps_system/CodeIgniter/application/controllers/manage_user.php
@@ -9,16 +9,40 @@ class Manage_user extends CI_Controller
     }
 	public function index()
 	{
-		$data['user_data'] = $this->manage_data->get_allusers();
-		$this->load->view('includes/header');
-		$this->load->view('system_views/manage_user', $data);
-		$this->load->view('includes/footer');
+		if($this->session->userdata('user_id') != null &&  $this->session->userdata('privilege') == 2)
+		{
+			$data['user_data'] = $this->manage_data->get_allusers();
+			$this->load->view('includes/header');
+			$this->load->view('system_views/manage_user', $data);
+			$this->load->view('includes/footer');
+		}
+		else
+		{
+			redirect('log_out');
+		}
 	}
+
+	//To delete user.
 	public function delete_user($remove_id)
 	{
-		$result = $this->manage_data->delete_user($remove_id);
-		$this->session->set_flashdata('success_msg', 'One row has been deleted.');
-		redirect('manage_user');
+		if($this->session->userdata('user_id') != null &&  $this->session->userdata('privilege') == 2)
+		{
+			$result = $this->manage_data->delete_user($remove_id);
+			if($result)
+			{
+				$this->session->set_flashdata('success_msg', 'One row has been deleted.');
+				redirect('manage_user');
+			}
+			else
+			{
+				$this->session->set_flashdata('error_msg', 'Sorry u can not delete this user.');
+				redirect('manage_user');
+			}
+		}
+		else
+		{
+			redirect('log_out');
+		}
 
 	}
 }
